Fix fs not showing current and parent directories (#29)
In previous Linux kernel versions (e.g., 5.8), "ls -la" would not show
"." and ".." in the root directory.

The rooot cause is, in Linux systems, there are different ways to
implement the "ls" command. e.g., Ubuntu systems use glibc, while
embedded systems use busybox. It is evident that the ext2/ext3/ext4
systems all skip inode number 0. The ext4 kernel document also states
that "There is no inode 0."

In the patch titled "patchwork - vfs: avoid creation of inode number 0
in get_next_ino," [1] new code was added to avoid providing the 0 inode.
In Linux v6.3, the corresponding patch was landed.

Fix solution:
To address the issue, we follow the ext4 rule and skip inode number 0.
When creating the image layout in the inode section, we skip the first
block and use the next inode (inode number 1) for the root inode. We also
modify "d_make_root" to use inode number 1 for the root inode.                                                                                                

[1] https://patchwork.kernel.org/project/linux-fsdevel/patch/1435245958-4507-1-git-send-email-cmaiolino@redhat.com/

Close #12