Fix errors raised by missing-prototypes
The recent upstream commit (0fcb708) in the Linux kernel globally
enforces function prototype checks. This means that it's crucial
for each function to have a declaration to validate that both the
caller and the callee expect the same argument types. Failure to
comply with this requirement may result in real bugs due to
mismatched prototypes. Therefore, in this commit, we need to add
function prototypes to the header files or declare functions with
static to ensure compatibility with these changes.