Fix file count with statfs
The 'f_files' member was incorrectly set to the count of inodes in use.
It should be set to the maximum number of files that the file system can
support. This incorrect calculation leads to unexpected behavior in
programs that use statfs(2), such as 'df'. These programs may print
negative values for the inodes used field, which is counterintuitive.