Add journal to write function
In the current design of simplefs, we can only access the extent
metadata and perform the 'journal' action on it. To write inode and
other metadata, we might need to emulate `ext4_dirty_inode`.
This method primarily uses the content of struct ext4_iloc to find the
correct location of the inode on disk and records it in the journal.
However, in our current situation, simply using `mark_inode_dirty` to
handle dirty inodes does not allow us to obtain the inode's location,
thus preventing us from recording it in the journal.

But currently, the idea is that we might create a new buffer head, use
`jbd2_journal_get_create_access` to put the dirty inode in it, and then
use `jbd2_journal_dirty_metadata` to write it to the journal.