Implement External Journal Device Initialization
This update introduces the following changes:

1. Added a function simplefs_parse_options in the fill super process
to obtain the external journal device and configure it using jbd2
related functions. This allows the file system to correctly initialize
and manage the external journal device.

2. Implemented additional functions in put_super to handle the journal
device during the unmount process. This ensures that the journal device
is properly managed and released when the file system is unmounted.

3. Added a make journal command in the Makefile. This new command allows
 users to create an external journal device image and mount it to
 simplefs, simplifying the setup and usage of the external journal.

Future Work:
Currently, the external journal device size is fixed at 8MB. To use a
different size, corresponding variables in the code need to be modified.
Additionally, support for an internal journal (inode journal) can be
added in the future to further improve the file system's capabilities.